step 4 refactor: utility folder to store sub functions
